add an sdk_test to make absolutely certain the adjustment for fingerprint calculation (when the gap between blocks is bigger than 32 bit rep) is correct (for reasonably sized files)

+++ src/filefingerprint.cpp
@@ -299,8 +299,9 @@ bool FileFingerprint::genfingerprint(InputStreamAccess *is, m_time_t cmtime, boo
                         size = -1;
                         return true;
                     }
-                    fullstep -= step;
+                    fullstep -= (uint64_t)step;
                 }
+
                 current += (offset - current);
 
                 if (!is->read(block, sizeof block))
